+++ src/App.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
-import { useOptionBgAnimationStore } from '@/store';
+import { useOptionAudioStore, useOptionBgAnimationStore } from '@/store';
 import TheHeader from '@/components/TheHeader.vue';
 import TheFooter from '@/components/TheFooter.vue';
 import OptionalComponent from '@/components/OptionalComponent.vue';
 
 const optionBgAnimation = useOptionBgAnimationStore();
+const optionAudio = useOptionAudioStore();
 </script>
 
 <template>
@@ -20,6 +21,13 @@ const optionBgAnimation = useOptionBgAnimationStore();
   <transition name="fade">
     <TheFooter v-visible="!optionBgAnimation.isPlaying"></TheFooter>
   </transition>
+  <OptionalComponent
+    v-if="optionAudio.canPlay"
+    comp="AudioPlayer"
+    :is-playing="optionAudio.isPlaying"
+    :volume="optionAudio.volume.value"
+    :name="optionAudio.selected.value"
+  ></OptionalComponent>
 </template>
 
 <style lang="scss">