+++ src/store/index.ts
@@ -2,7 +2,9 @@ import { createPinia } from 'pinia';
 
 export * from './use-entries-store';
 export * from './use-selected-entries-store';
+
 export * from './use-options-store';
+export * from './use-option-audio-store';
 export * from './use-option-bg-animation-store';
 
 export default createPinia();